pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "alphapy-pro"
dynamic = ["version"]
description = "AlphaPy Pro: A Machine Learning Pipeline for Speculators"
readme = "README.md"
license = "Apache-2.0"
authors = [
    {name = "Robert D. Scott II", email = "scottfree.analytics@scottfreellc.com"},
    {name = "Mark Conway", email = "scottfree.analytics@scottfreellc.com"},
]
maintainers = [
    {name = "ScottFree LLC", email = "scottfree.analytics@scottfreellc.com"},
]
keywords = ["machine learning", "trading", "finance", "prediction", "scikit-learn", "pandas"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Operating System :: OS Independent",
]
requires-python = ">=3.12"
dependencies = [
    "arrow>=0.13",
    "bokeh>=1.3",
    "category_encoders>=2.1",
    "imbalanced-learn>=0.5",
    "ipython>=7.2",
    "lofo-importance>=0.3.4",
    "matplotlib>=3.0",
    "numpy>=1.20",
    "pandas>=1.3",
    "pandas-datareader>=0.10.0",
    "polygon-api-client>=1.13.4",
    "pyyaml>=5.0",
    "scikit-learn>=1.0",
    "scipy>=1.7",
    "seaborn>=0.11",
    "setuptools>=61.0",
    "statsmodels>=0.13.0",
    "venn-abers>=1.4.1",
    "xgboost>=3.0.4",
    "catboost>=1.2",
    "yfinance>=0.1.59",
    "lightgbm>=4.0.0",
    "optuna>=4.0.0",
    "optuna-integration[sklearn]>=4.0.0",
    # Trading agent dependencies
    "alpaca-py>=0.21.0",
    "anthropic>=0.25.0",
    # High-performance data & indicators
    "polars>=1.0.0",
    "talipp>=2.7.0",
    "pydantic>=2.0.0",
    "quantstats>=0.0.77",
    "pyarrow>=22.0.0",
    # Vectorized backtesting
    "vectorbt>=0.26.0",
    # Alfi markets app — web frameworks
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "streamlit",
    "reflex>=0.6.0",
    # Alfi markets app — async & networking
    "aiohttp>=3.8.0",
    "websockets",
    "requests>=2.28.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.6",
    # Alfi markets app — market data providers
    "massive",
    "finnhub-python",
    "finviz",
    # Alfi markets app — visualization & indicators
    "plotly>=5.0.0",
    "pandas-ta",
    # Alfi markets app — LLM / AI
    "openai",
    "litellm>=1.81.13",
    # Alfi markets app — portfolio optimization
    "skfolio>=0.4.0",
    "joblib>=1.3.0",
    # Alfi markets app — pattern discovery (Phase 7)
    "stumpy>=1.12.0",
    "tsfresh>=0.20.0",
    "pycatch22>=0.4.0",
    "deap>=1.4.0",
    # Alfi markets app — utilities & publishing
    "python-dotenv>=1.0.0",
    "jinja2>=3.0.0",
    "schedule>=1.2.0",
    "lxml>=4.9.0",
    "html5lib>=1.1",
    "python-frontmatter>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=22.0",
    "isort>=5.0",
    "flake8>=4.0",
    "pre-commit>=2.0",
    "build>=0.8",
    "twine>=4.0",
]
docs = [
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0",
    "sphinx-rtd-dark-mode>=1.0",
]
test = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "pytest-xdist>=2.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-timeout>=2.1.0",
    "freezegun>=1.2.0",
]

[project.urls]
Homepage = "https://github.com/ScottFreeLLC/alphapy-pro"
Documentation = "https://scottfreellc.github.io/alphapy-pro/"
Repository = "https://github.com/ScottFreeLLC/alphapy-pro.git"
"Bug Tracker" = "https://github.com/ScottFreeLLC/alphapy-pro/issues"

[project.scripts]
alphapy = "alphapy.alphapy_main:main"
mflow = "alphapy.mflow_main:main"
scott = "scripts.run_agent:main"

[tool.hatch.version]
path = "alphapy/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["alphapy", "agent", "app", "scripts"]

[tool.uv]
dev-dependencies = [
    "pytest>=6.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["alphapy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-v",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "asyncio: marks tests as async (pytest-asyncio)",
]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::FutureWarning",
]

# Coverage settings (used when pytest-cov is installed)
# Run with: pytest --cov=alphapy --cov-report=term-missing
[tool.pytest_cov]
addopts = [
    "--cov=alphapy",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.coverage.run]
source = ["alphapy", "agent"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
fail_under = 20
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "sklearn.*", 
    "pandas.*", 
    "numpy.*", 
    "matplotlib.*", 
    "seaborn.*",
    "bokeh.*",
    "yfinance.*",
    "polygon.*",
]
ignore_missing_imports = true
